+++ cdk/drupal/drupal_stack.py
@@ -30,12 +30,28 @@ class DrupalStack(core.Stack):
 
         # TODO: Encryption
         # https://github.com/aws/aws-cdk/blob/v1.36.1/packages/@aws-cdk/aws-codepipeline/lib/pipeline.ts#L225-L244
-        artifact_bucket = aws_s3.Bucket(
+        pipeline_artifact_bucket_name_param = core.CfnParameter(
+            self,
+            "PipelineArtifactBucketName",
+            default=""
+        )
+        pipeline_artifact_bucket_name_not_exists_condition = core.CfnCondition(
+            self,
+            "PipelineArtifactBucketNameNotExists",
+            expression=core.Fn.condition_equals(pipeline_artifact_bucket_name_param.value, "")
+        )
+        pipeline_artifact_bucket_name_exists_condition = core.CfnCondition(
+            self,
+            "PipelineArtifactBucketNameExists",
+            expression=core.Fn.condition_not(core.Fn.condition_equals(pipeline_artifact_bucket_name_param.value, ""))
+        )
+        pipeline_artifact_bucket = aws_s3.CfnBucket(
             self,
             "ArtifactBucket",
-            access_control=aws_s3.BucketAccessControl.PRIVATE,
-            block_public_access=aws_s3.BlockPublicAccess.BLOCK_ALL
+            access_control="Private",
+            public_access_block_configuration=aws_s3.BlockPublicAccess.BLOCK_ALL
         )
+        pipeline_artifact_bucket.cfn_options.condition=pipeline_artifact_bucket_name_not_exists_condition
 
         source_artifact_s3_bucket_param = core.CfnParameter(
             self,
@@ -805,7 +821,7 @@ class DrupalStack(core.Stack):
                             resources=[
                                 "arn:{}:s3:::{}/*".format(
                                     core.Aws.PARTITION,
-                                    artifact_bucket.bucket_name
+                                    pipeline_artifact_bucket.ref
                                 )
                             ]
                         )
@@ -1114,7 +1130,7 @@ class DrupalStack(core.Stack):
                             resources=[
                                 "arn:{}:s3:::{}/*".format(
                                     core.Aws.PARTITION,
-                                    artifact_bucket.bucket_name
+                                    pipeline_artifact_bucket.ref
                                 )
                             ]
                         )
@@ -1146,7 +1162,7 @@ class DrupalStack(core.Stack):
                             resources=[
                                 "arn:{}:s3:::{}/*".format(
                                     core.Aws.PARTITION,
-                                    artifact_bucket.bucket_name
+                                    pipeline_artifact_bucket.ref
                                 )
                             ]
                         )
@@ -1194,7 +1210,7 @@ class DrupalStack(core.Stack):
             self,
             "Pipeline",
             artifact_store=aws_codepipeline.CfnPipeline.ArtifactStoreProperty(
-                location=artifact_bucket.bucket_name,
+                location=pipeline_artifact_bucket.ref,
                 type='S3'
             ),
             role_arn=pipeline_role.role_arn,